+++ src/main/java/com/ezen/valuefinder/controller/MainController.java
@@ -11,28 +11,10 @@ public class MainController {
         return "index";
     }
     
-    @GetMapping(value = "/inquiry")
-    public String inquiry() {
-        return "inquirys/inquiry";
-    }
     
-    @GetMapping(value = "/login2")
-    public String login() {
-        return "member/login";
-    }
-
-    @GetMapping(value = "/inquiryView")
-    public String upinquiry() {
-        return "inquirys/inquiryView";
-    }
     
-    @GetMapping(value = "/report")
-    public String report() {
-        return "inquirys/report";
-    }
+   
+    
 
-    @GetMapping(value = "/publicbidDetails")
-    public String publicbidDetails() {
-    	return "details/publicbidDetails";
-    }
+    
 }
